backend/requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
google-api-python-client>=2.150.0,<3
google-auth>=2.35.0,<3
python-dotenv>=1.0.0,<2
httpx>=0.27.0,<1
pytest>=8.0.0,<9
